+++ packages/stk/stk_unit_tests/stk_mesh/ngp/ngpMultiStateFieldTests.cpp
@@ -40,7 +40,7 @@
 #include <stk_mesh/base/MetaData.hpp>
 #include <stk_mesh/base/BulkData.hpp>
 #include <stk_mesh/base/Field.hpp>
-#include <stk_mesh/base/FieldBLAS.hpp>
+#include <stk_mesh/base/NgpFieldBLAS.hpp>
 #include <stk_mesh/base/Bucket.hpp>
 #include <stk_mesh/base/Entity.hpp>
 #include <stk_mesh/base/GetEntities.hpp>
@@ -200,8 +200,8 @@ NGP_TEST_F(NgpMultiStateFieldTest, multistateField_rotateDeviceStates_syncStates
 
   const double valueNew = 44.4;
   const double valueOld = 22.2;
-  stk::mesh::field_fill(valueNew, get_field_new());
-  stk::mesh::field_fill(valueOld, get_field_old());
+  stk::mesh::field_fill(valueNew, get_field_new(), stk::ngp::HostExecSpace());
+  stk::mesh::field_fill(valueOld, get_field_old(), stk::ngp::HostExecSpace());
 
   stk::mesh::NgpMesh& ngpMesh = stk::mesh::get_updated_ngp_mesh(get_bulk());
   stk::mesh::NgpField<double>& ngpFieldNew = stk::mesh::get_updated_ngp_field<double>(get_field_new());
@@ -230,8 +230,8 @@ NGP_TEST_F(NgpMultiStateFieldTest, multistateField_copyHasCorrectDataAfterStateR
 
   const double valueNew = 44.4;
   const double valueOld = 22.2;
-  stk::mesh::field_fill(valueNew, get_field_new());
-  stk::mesh::field_fill(valueOld, get_field_old());
+  stk::mesh::field_fill(valueNew, get_field_new(), stk::ngp::HostExecSpace());
+  stk::mesh::field_fill(valueOld, get_field_old(), stk::ngp::HostExecSpace());
 
   stk::mesh::NgpMesh& ngpMesh = stk::mesh::get_updated_ngp_mesh(get_bulk());
   stk::mesh::NgpField<double>& ngpFieldNew = stk::mesh::get_updated_ngp_field<double>(get_field_new());
@@ -260,8 +260,8 @@ NGP_TEST_F(NgpMultiStateFieldTest, multistateField_copyHasWrongDataAfterDeviceSt
 
   const double valueNew = 44.4;
   const double valueOld = 22.2;
-  stk::mesh::field_fill(valueNew, get_field_new());
-  stk::mesh::field_fill(valueOld, get_field_old());
+  stk::mesh::field_fill(valueNew, get_field_new(), stk::ngp::HostExecSpace());
+  stk::mesh::field_fill(valueOld, get_field_old(), stk::ngp::HostExecSpace());
 
   stk::mesh::NgpMesh& ngpMesh = stk::mesh::get_updated_ngp_mesh(get_bulk());
   stk::mesh::NgpField<double>& ngpFieldNew = stk::mesh::get_updated_ngp_field<double>(get_field_new());
@@ -293,8 +293,8 @@ NGP_TEST_F(NgpMultiStateFieldTest, persistentDeviceField_hasCorrectDataAfterStat
 
   const double valueNew = 44.4;
   const double valueOld = 22.2;
-  stk::mesh::field_fill(valueNew, get_field_new());
-  stk::mesh::field_fill(valueOld, get_field_old());
+  stk::mesh::field_fill(valueNew, get_field_new(), stk::ngp::HostExecSpace());
+  stk::mesh::field_fill(valueOld, get_field_old(), stk::ngp::HostExecSpace());
 
   stk::mesh::NgpMesh& ngpMesh = stk::mesh::get_updated_ngp_mesh(get_bulk());
   stk::mesh::NgpField<double>& ngpFieldNew = stk::mesh::get_updated_ngp_field<double>(get_field_new());
@@ -322,8 +322,8 @@ NGP_TEST_F(NgpMultiStateFieldTest, persistentDeviceField_hasWrongDataAfterDevice
 
   const double valueNew = 44.4;
   const double valueOld = 22.2;
-  stk::mesh::field_fill(valueNew, get_field_new());
-  stk::mesh::field_fill(valueOld, get_field_old());
+  stk::mesh::field_fill(valueNew, get_field_new(), stk::ngp::HostExecSpace());
+  stk::mesh::field_fill(valueOld, get_field_old(), stk::ngp::HostExecSpace());
 
   stk::mesh::NgpMesh& ngpMesh = stk::mesh::get_updated_ngp_mesh(get_bulk());
   stk::mesh::NgpField<double>& ngpFieldNew = stk::mesh::get_updated_ngp_field<double>(get_field_new());